requirements.txt
numpy
pyzbar
Pillow
opencv-python
